cifs: changes to support multichannel during channel  reconnect
The reconnect scenario when using a multichannel mount had a
few issues:

1. A tcp connection on binding channel did not have a way to
get to the primary connection or smb channel easily.
I replaced is_channel field with primary_server, which is a
referenced pointer to the primary channel tcp connection.
This also allows us to get to the smb session for the channel.

2. In case of cifs_reconnect, we needed a way to identify the
channels under the smb session which are in reconnect, so that
the traffic to other channels can continue. So I replaced
the bool need_reconnect with a bitmask identifying all the
channels that are currently needing reconnection and called it
chans_need_reconnect. When a channel needs reconnection, the bit
corresponding to the index of the server in ses->chans is used
to set this bitmask. Checking if no channels or all the channels
need reconnect then becomes very easy.

3. Mark tcon->need_reconnect only when all the channels underneath
need to reconnect.

Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
Signed-off-by: Steve French <stfrench@microsoft.com>